+++ modules/appnexusBidAdapter.js
@@ -1,5 +1,4 @@
 import {
-  chunk,
   createTrackPixelHtml,
   deepAccess,
   deepClone,
@@ -37,6 +36,7 @@ import {
 } from '../libraries/appnexusUtils/anKeywords.js';
 import {convertCamelToUnderscore, fill} from '../libraries/appnexusUtils/anUtils.js';
 import {convertTypes} from '../libraries/transformParamsUtils/convertTypes.js';
+import {chunk} from '../libraries/chunk/chunk.js';
 
 const BIDDER_CODE = 'appnexus';
 const URL = 'https://ib.adnxs.com/ut/v3/prebid';
